+++ src/odb/src/db/dbTechLayerMinCutRule.cpp
@@ -109,7 +109,7 @@ dbIStream& operator>>(dbIStream& stream, _dbTechLayerMinCutRule& obj)
     dbDatabase* db = (dbDatabase*) obj.getDatabase();
     int area;
     stream >> area;
-    obj.area_ = area * db->getDbuPerMicron();
+    obj.area_ = static_cast<int64_t>(area) * db->getDbuPerMicron();
   }
   // User Code End >>area_
   stream >> obj.area_within_dist_;
